Add Speed, Particles per second and loop options in inspector

+++ Factory Engin3D/Timer.cpp
@@ -38,14 +38,14 @@ Uint32 Timer::Read()
 	}
 }
 
-Uint32 Timer::ReadSec()
+float Timer::ReadSec()
 {
 	if (running == true)
 	{
-		return (SDL_GetTicks() - started_at)/1000;
+		return (SDL_GetTicks() - started_at)/1000.0f;
 	}
 	else
 	{
-		return (stopped_at - started_at)/1000;
+		return (stopped_at - started_at)/1000.0f;
 	}
 }
